% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/napr.R
\name{distance}
\alias{distance}
\title{Calculate distance between lat/long pairs.}
\usage{
distance(lat1, lat2, lon1, lon2)
}
\arguments{
\item{lat1}{a list of lat/long values for location 1}

\item{lat2}{a list of lat/long values for location 2}

\item{lon1}{a list of lat/long values for location 1}

\item{lon2}{a list of lat/long values for location 2}
}
\value{
distance from loc1 to loc2.
}
\description{
Calculate distance between lat/long pairs.
}

