% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ghelp.R
\name{ghelp}
\alias{ghelp}
\title{ghelp, look for function use in github source}
\usage{
ghelp(topic, in_cran = TRUE)
}
\arguments{
\item{topic,}{string topic to search for}
}
\value{
html to search for
}
\description{
ghelp, look for function use in github source
}

