% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ghelp.R
\name{githelp}
\alias{githelp}
\title{Look for function use in github source.}
\usage{
githelp(topic, in_cran = TRUE)
}
\arguments{
\item{topic,}{string topic to search for.}

\item{in_cran,}{whether the cran limiter should be applied. Default TRUE.}
}
\value{
html object containing source code with the topic.
}
\description{
Search github source for topics to find examples
of use of the topic. Based on:
https://gist.github.com/hrbrmstr/32e9c140129d7d51db52.
}

