% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/napr.R
\name{\%fin\%}
\alias{\%fin\%}
\title{Fast %in% operator using fastmatch.}
\usage{
x \%fin\% table
}
\arguments{
\item{x}{a value.}

\item{table}{the lookup table.}
}
\value{
a logical.
}
\description{
Fast %in% operator using fastmatch.
}

