% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/napr.R
\name{is.leapYear}
\alias{is.leapYear}
\title{Tests if year is a leap year.}
\usage{
is.leapYear(yr)
}
\arguments{
\item{year}{an integer}
}
\value{
boolean
}
\description{
Tests if year is a leap year.
}
\examples{
is.leapYear(2000) #TRUE
is.leapYear(2003) #FALSE
}

