% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{napr-package}
\alias{napr-package}
\title{napr: Scraped access to all NAPs published}
\description{
A package providing access to National Adaptation Plans (NAPs) from the UNFCCC NAP Central website. Contains pre-processed text data from NAP documents, regularly updated via GitHub Actions, with functions to access the latest versions when needed.
}
\author{
\strong{Maintainer}: Nils RC8rstad \email{202302046@post.au.dk}

}
\keyword{internal}
