% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napr.file.reader.20160510.R, R/napr.file.reader.20161129.R
\name{read.upload.file}
\alias{read.upload.file}
\title{read .brainage_upload_file}
\usage{
read.upload.file(input)

read.upload.file(input)
}
\arguments{
\item{input}{a tar archive containing lh.thickness.fwhm0.fsaverage4.mgh and rh.thickness.fwhm0.fsaverage4.mgh for a subject}

\item{file.name}{The name of a file that is the aseg.stats, lh.aparc and rh.aparc files cat'ed together}
}
\value{
a matrix with the numbers laid out so the predictive model can read them

a matrix with the numbers laid out so the predictive model can read them
}
\description{
read .brainage_upload_file

read .brainage_upload_file
}

