% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.age.multiple.models.R
\name{report.age.multiple.models}
\alias{report.age.multiple.models}
\title{feed freesurfer stats into model to predict age}
\usage{
report.age.multiple.models(file, model.list = list(rvmm.full.20161228.model,
  gausspr.full.20161228.model))
}
\arguments{
\item{file}{uploaded tar archive}

\item{model.list}{a list of predictive models generated using kernlab library}
}
\value{
predicted age
}
\description{
feed freesurfer stats into model to predict age
}

