% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.age.R
\name{report.age}
\alias{report.age}
\title{feed freesurfer stats into model to predict age}
\usage{
report.age(file, model = rvmm.20161104)
}
\arguments{
\item{file}{concatenated aseg.stats, lh.aparc and rh.aparc files}

\item{model}{a predictive model generated using kernlab library}
}
\value{
predicted age
}
\description{
feed freesurfer stats into model to predict age
}

