% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/napr.R
\name{round_any}
\alias{round_any}
\title{Rounds to arbitrary number.}
\usage{
round_any(n, r)
}
\arguments{
\item{n}{number to round.}

\item{r}{value to round to.}
}
\value{
numeric.
}
\description{
Rounds to arbitrary number.
}
\examples{
round_any(1.32, .5)
#> 1.5
}

