% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rvmm.fsavg4}
\alias{rvmm.fsavg4}
\title{Age prediction models
generated using rvmm part of kernlab library}
\format{kernlab prediction model}
\source{
a bunch of control MRI datasets
}
\usage{
rvmm.fsavg4
}
\description{
Age prediction models
generated using rvmm part of kernlab library
}
\keyword{datasets}

