% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rvmm}
\alias{rvmm}
\title{Age prediction model
I can't remember all the datasets that went into it but I'll fix it
generated using rvmm part of kernlab library}
\format{kernlab prediction model}
\source{
a bunch of control MRI datasets
}
\usage{
rvmm
}
\description{
Age prediction model
I can't remember all the datasets that went into it but I'll fix it
generated using rvmm part of kernlab library
}
\keyword{datasets}

