% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napr.file.reader.20160510.R
\name{read.upload.file}
\alias{read.upload.file}
\title{read .brainage_upload_file}
\usage{
read.upload.file(file.name)
}
\arguments{
\item{file.name}{The name of a file that is the aseg.stats, lh.aparc and rh.aparc files cat'ed together}
}
\value{
a matrix with the numbers laid out so the predictive model can read them
}
\description{
read .brainage_upload_file
}

